renamed solutionSearch file and its exported function
